+++ src/colour_matrix.h
@@ -21,9 +21,9 @@ namespace libdcp {
 
 namespace colour_matrix {
 
-extern float const xyz_to_rgb[3][3];
-extern float const srgb_to_xyz[3][3];
-extern float const rec709_to_xyz[3][3];
+extern double const xyz_to_rgb[3][3];
+extern double const srgb_to_xyz[3][3];
+extern double const rec709_to_xyz[3][3];
 
 }
